refactor Customer filter
