+++ app/Filters/ApiFilter.php
@@ -1,30 +1,14 @@
 <?php
 
-namespace App\Services\V1;
+namespace App\Filters;
 
-class CustomerQuery
+class ApiFilter
 {
-    protected $safeParams = [
-        'name'       => ['eq'],
-        'type'       => ['eq'],
-        'email'      => ['eq'],
-        'address'    => ['eq'],
-        'city'       => ['eq'],
-        'state'      => ['eq'],
-        'postalCode' => ['eq', 'gt', 'lt']
-    ];
+    protected $safeParams = [];
 
-    protected $columnMap = [
-        'postalCode' => 'postal_code'
-    ];
+    protected $columnMap = [];
 
-    protected $operatorMap = [
-        'eq' => '=',
-        'gt' => '>',
-        'gte' => '>=',
-        'lt' => '<',
-        'lte' => '<='
-    ];
+    protected $operatorMap = [];
 
     public function transform($request)
     {
