+++ app/Filters/V1/CustomersFilter.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace App\Filters\V1;
+
+use App\Filters\ApiFilter;
+
+class CustomersFilter extends ApiFilter
+{
+    protected $safeParams = [
+        'name'       => ['eq'],
+        'type'       => ['eq'],
+        'email'      => ['eq'],
+        'address'    => ['eq'],
+        'city'       => ['eq'],
+        'state'      => ['eq'],
+        'postalCode' => ['eq', 'gt', 'lt']
+    ];
+
+    protected $columnMap = [
+        'postalCode' => 'postal_code'
+    ];
+
+    protected $operatorMap = [
+        'eq' => '=',
+        'gt' => '>',
+        'gte' => '>=',
+        'lt' => '<',
+        'lte' => '<='
+    ];
+}
